requirements.txt
google_api_python_client==2.166.0
google_auth_oauthlib==1.2.1
protobuf==6.30.2
